gotta update LDLib and dump or update registrate

+++ src/main/java/com/gregtechceu/gtceu/api/data/chemical/material/properties/FluidPipeProperties.java
@@ -22,7 +22,7 @@ public class FluidPipeProperties implements IMaterialProperty<FluidPipePropertie
     public static final int MAX_PIPE_CHANNELS = 9;
 
     @Getter @Setter
-    private long throughput;
+    private int throughput;
     @Getter @Setter
     private int channels;
     @Getter @Setter
@@ -37,7 +37,7 @@ public class FluidPipeProperties implements IMaterialProperty<FluidPipePropertie
     private final Object2BooleanMap<FluidAttribute> containmentPredicate = new Object2BooleanOpenHashMap<>();
 
 
-    public FluidPipeProperties(int maxFluidTemperature, long throughput, boolean gasProof, boolean acidProof, boolean cryoProof, boolean plasmaProof, int channels) {
+    public FluidPipeProperties(int maxFluidTemperature, int throughput, boolean gasProof, boolean acidProof, boolean cryoProof, boolean plasmaProof, int channels) {
         this.maxFluidTemperature = maxFluidTemperature;
         this.throughput = throughput;
         this.gasProof = gasProof;
@@ -50,7 +50,7 @@ public class FluidPipeProperties implements IMaterialProperty<FluidPipePropertie
     /**
      * Default property constructor.
      */
-    public FluidPipeProperties(int maxFluidTemperature, long throughput, boolean gasProof, boolean acidProof, boolean cryoProof, boolean plasmaProof) {
+    public FluidPipeProperties(int maxFluidTemperature, int throughput, boolean gasProof, boolean acidProof, boolean cryoProof, boolean plasmaProof) {
         this(maxFluidTemperature, throughput, gasProof, acidProof, cryoProof, plasmaProof, 1);
     }
